refactor: extract InitFileExtractor to hilla-plugin-base
hilla-plugin-base contains base/shared classes
needed by hilla plugins in general.
